Move xREL release DTO to shared, per project convention
XrelReleaseListItem (and its ext_info shape, XrelExtInfo) was defined
separately in server/xrel.ts and duplicated by hand in
client/src/pages/xrel-releases.tsx, so the two could silently drift.
Moved the transport-level DTO into shared/xrel-types.ts and imported
it from both sides; server/xrel.ts still re-exports it for existing
callers. Page-only fields (isWanted, libraryStatus, gameId,
matchCandidate) stay local to the client as an extension of the
shared type.
